mp_heatmap

mp_heatmap

mp_heatmap(
  locality,
  cve_edo,
  geocoded_datasets,
  status_caso,
  zoom,
  week,
  alpha,
  map_type
)

Arguments

locality

is the locality locality target.

cve_edo

is a string of the state id.

geocoded_datasets

is geocoded dataset.

status_caso

It is a numerical value to select the Status of the case. There are three options, 1 for the probable cases database, 2 for the confirmed cases database, and 3 for the discarded cases database.

zoom

map zoom; an integer from 3 (continent) to 21 (building), default value 10 (city).

week

It is epidemiological week.

alpha

is a numerical parameter that controls the transparency of the heatmap. Values range from 0 to 1, where 0 is completely transparent and 1 does not make the heat map transparent.

map_type

character string providing google map theme. options available are "terrain", "satellite", "roadmap", and "hybrid"

Value

ggplot object

Author

Felipe Antonio Dzul Manzanilla felipe.dzul.m@gmail.com